code linting: rearrange spec files
