+++ client/src/tests/actions/UserActions/updateUserSpec.js
@@ -1,28 +1,29 @@
 import chai from 'chai';
 import sinon from 'sinon';
-import MockHttpClient from '../MockHttpClient';
-import MockHttpErrorClient from '../MockHttpErrorClient';
-import * as ActionTypes from '../../constants';
-import { deleteUser } from '../../actions/UserActions';
+import MockHttpClient from '../../MockHttpClient';
+import MockHttpErrorClient from '../../MockHttpErrorClient';
+import * as ActionTypes from '../../../constants';
+import { updateUser } from '../../../actions/UserActions';
 
 const expect = chai.expect;
 
-describe('deleteUser', () => {
+describe('updateUser', () => {
   it('should dispatch its success actions correctly', () => {
     const token = 'RANDOM_TOKEN';
     const targetUserId = 12;
-    const deleteUserActions = deleteUser(token, targetUserId);
+    const updateInfo = { firstName: 'Crawford' };
+    const updateUserActions = updateUser(token, targetUserId, updateInfo);
     const spy = sinon.spy();
-    deleteUserActions(spy, MockHttpClient);
+    updateUserActions(spy, MockHttpClient);
     expect(spy.calledTwice).to.equal(true);
     expect(spy.calledWith(
       {
-        type: ActionTypes.DELETE_USER_PENDING
+        type: ActionTypes.UPDATE_USER_PENDING
       }
     )).to.equal(true);
     expect(spy.calledWith(
       {
-        type: ActionTypes.DELETE_USER_FULFILLED,
+        type: ActionTypes.UPDATE_USER_FULFILLED,
         payload: {
           message: 'Request successful.'
         }
@@ -33,18 +34,19 @@ describe('deleteUser', () => {
   it('should dispatch its failure actions correctly', () => {
     const token = 'RANDOM_TOKEN';
     const targetUserId = 12;
-    const deleteUserActions = deleteUser(token, targetUserId);
+    const updateInfo = { firstName: 'Crawford' };
+    const updateUserActions = updateUser(token, targetUserId, updateInfo);
     const spy = sinon.spy();
-    deleteUserActions(spy, MockHttpErrorClient);
+    updateUserActions(spy, MockHttpErrorClient);
     expect(spy.calledTwice).to.equal(true);
     expect(spy.calledWith(
       {
-        type: ActionTypes.DELETE_USER_PENDING
+        type: ActionTypes.UPDATE_USER_PENDING
       }
     )).to.equal(true);
     expect(spy.calledWith(
       {
-        type: ActionTypes.DELETE_USER_REJECTED,
+        type: ActionTypes.UPDATE_USER_REJECTED,
         payload: {
           message: 'Request failed.'
         }
