+++ package.json
@@ -17,9 +17,9 @@
     "start": "SEQUELIZE_LOGGING=false ./node_modules/.bin/babel-node server.js",
     "start:watch": "./node_modules/.bin/nodemon server.js --exec babel-node --presets es2015,react,node6",
     "test": "npm run test:server",
-    "test:client": "./node_modules/.bin/mocha --require babel-register ./client/src/tests/setup.js ./client/src/tests/*/*[sS]pec.js",
+    "test:client": "./node_modules/.bin/mocha --require babel-register ./client/src/tests/setup.js ./client/src/tests/*/*[sS]pec.js ./client/src/tests/*/*/*[sS]pec.js",
     "test:nyc": "npm run test:nyc:client && npm run test:nyc:server",
-    "test:nyc:client": "SEQUELIZE_LOGGING=false ./node_modules/.bin/nyc --reporter=html --reporter=text ./node_modules/.bin/mocha --require babel-register ./client/src/tests/setup.js ./client/src/tests/*/*[sS]pec.js",
+    "test:nyc:client": "SEQUELIZE_LOGGING=false ./node_modules/.bin/nyc --reporter=html --reporter=text ./node_modules/.bin/mocha --require babel-register ./client/src/tests/setup.js ./client/src/tests/*/*[sS]pec.js ./client/src/tests/*/*/*[sS]pec.js",
     "test:nyc:server": "SEQUELIZE_LOGGING=false ./node_modules/.bin/nyc --reporter=html --reporter=text ./node_modules/.bin/mocha --timeout 5000 --require babel-register ./server/tests/*/*[sS]pec.js",
     "test:server": "SEQUELIZE_LOGGING=false ./node_modules/.bin/mocha --timeout 5000 --require babel-register ./server/tests/*/*[sS]pec.js",
     "test:travis": "NODE_ENV=test npm run test:server"
